requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pyodbc
afip.py==1.1.2
python-jose[cryptography]==3.3.0
bcrypt
python-dotenv==1.0.1
python-multipart==0.0.9
pywebview
pyinstaller
